blue review changes
