+++ ray_on_golem/server/cluster/__init__.py
@@ -1,3 +1,9 @@
 from ray_on_golem.server.cluster.cluster import Cluster
+from ray_on_golem.server.cluster.nodes import ClusterNode, HeadClusterNode, WorkerClusterNode
 
-__all__ = ("Cluster",)
+__all__ = (
+    "Cluster",
+    "ClusterNode",
+    "HeadClusterNode",
+    "WorkerClusterNode",
+)
