+++ ray_on_golem/server/main.py
@@ -14,9 +14,7 @@ from golem.utils.asyncio import create_task_with_logging, ensure_cancelled_many
 from ray_on_golem.cli import with_datadir
 from ray_on_golem.reputation.updater import ReputationUpdater
 from ray_on_golem.server.middlewares import error_middleware, trace_id_middleware
-from ray_on_golem.server.services import YagnaService
-from ray_on_golem.server.services.new_golem import GolemService
-from ray_on_golem.server.services.new_ray import RayService
+from ray_on_golem.server.services import GolemService, RayService, YagnaService
 from ray_on_golem.server.settings import (
     RAY_ON_GOLEM_SHUTDOWN_CONNECTIONS_TIMEOUT,
     WEBSOCAT_PATH,
