+++ ray_on_golem/server/services/__init__.py
@@ -1,11 +1,19 @@
-from ray_on_golem.server.services.golem import get_manifest
-from ray_on_golem.server.services.new_golem import GolemService
-from ray_on_golem.server.services.new_ray import RayService
+from ray_on_golem.server.services.golem import (
+    DriverListAllocationPaymentManager,
+    GolemService,
+    ManagerStack,
+    get_manifest,
+)
+from ray_on_golem.server.services.ray import RayService
 from ray_on_golem.server.services.yagna import YagnaService
 
 __all__ = (
     "GolemService",
     "RayService",
     "YagnaService",
+    "get_manifest",
+    "GolemService",
+    "DriverListAllocationPaymentManager",
+    "ManagerStack",
     "get_manifest",
 )
