+++ ray_on_golem/server/services/golem/__init__.py
@@ -1,7 +1,13 @@
-from ray_on_golem.server.services.golem.golem import GolemService
+from ray_on_golem.server.services.golem.golem import (
+    DriverListAllocationPaymentManager,
+    GolemService,
+)
+from ray_on_golem.server.services.golem.manager_stack import ManagerStack
 from ray_on_golem.server.services.golem.manifest import get_manifest
 
 __all__ = (
     "GolemService",
+    "DriverListAllocationPaymentManager",
+    "ManagerStack",
     "get_manifest",
 )
