+++ ray_on_golem/server/services/golem/helpers/__init__.py
@@ -1,0 +1,7 @@
+from ray_on_golem.server.services.golem.helpers.demand_config import DemandConfigHelper
+from ray_on_golem.server.services.golem.helpers.manager_stack import ManagerStackNodeConfigHelper
+
+__all__ = (
+    "DemandConfigHelper",
+    "ManagerStackNodeConfigHelper",
+)
